Move ads to new layout
