+++ benchmarks/eko/benchmark_ad.py
@@ -2,7 +2,7 @@
 import numpy as np
 import pytest
 
-import ekore.anomalous_dimensions.as2 as ad_as2
+import ekore.anomalous_dimensions.unpolarized.space_like.as2 as ad_as2
 import ekore.harmonics as h
 from eko.constants import CA, CF, TR
