+++ src/eko/evolution_operator/__init__.py
@@ -13,7 +13,7 @@ import numba as nb
 import numpy as np
 from scipy import integrate
 
-from ekore import anomalous_dimensions as ad
+import ekore.anomalous_dimensions.unpolarized.space_like as ad
 from .. import basis_rotation as br
 from .. import interpolation, mellin
 from .. import scale_variations as sv
