+++ src/ekore/anomalous_dimensions/unpolarized/space_like/as4/gnsm.py
@@ -4,8 +4,8 @@
 import numba as nb
 
 from eko.constants import CF
-from ...harmonics.constants import zeta3
-from ...harmonics.log_functions import lm11m1, lm12m1, lm13m1
+from .....harmonics.constants import zeta3
+from .....harmonics.log_functions import lm11m1, lm12m1, lm13m1
 
 
 @nb.njit(cache=True)
