+++ tests/ekore/anomalous_dimensions/unpolarized/space_like/test_init.py
@@ -6,8 +6,9 @@ from numpy.testing import assert_allclose, assert_almost_equal, assert_raises
 
 from ekore import anomalous_dimensions as ad
 from eko import basis_rotation as br
-from ekore.anomalous_dimensions import as1 as ad_as1
-from ekore.anomalous_dimensions import harmonics
+import ekore.anomalous_dimensions.unpolarized.space_like as ad_us
+from ekore.anomalous_dimensions.unpolarized.space_like import as1 as ad_as1
+from ekore import harmonics
 
 NF = 5
 
@@ -41,7 +42,7 @@ def test_eigensystem_gamma_singlet_projectors_EV():
             # NNLO and N3LO too big numbers,
             # ignore Runtime Warnings
             warnings.simplefilter("ignore", RuntimeWarning)
-            for gamma_S in ad.gamma_singlet(o, N, nf):
+            for gamma_S in ad_us.gamma_singlet(o, N, nf):
                 _exp, l_p, l_m, e_p, e_m = ad.exp_singlet(gamma_S)
                 # projectors behave as P_a . P_b = delta_ab P_a
                 assert_allclose(np.dot(e_p, e_p), e_p)
@@ -56,40 +57,40 @@ def test_gamma_ns():
     nf = 3
     # LO
     assert_almost_equal(
-        ad.gamma_ns((3, 0), br.non_singlet_pids_map["ns+"], 1, nf)[0], 0.0
+        ad_us.gamma_ns((3, 0), br.non_singlet_pids_map["ns+"], 1, nf)[0], 0.0
     )
     # NLO
     assert_allclose(
-        ad.gamma_ns((2, 0), br.non_singlet_pids_map["ns-"], 1, nf),
+        ad_us.gamma_ns((2, 0), br.non_singlet_pids_map["ns-"], 1, nf),
         np.zeros(2),
         atol=2e-6,
     )
     # NNLO
     assert_allclose(
-        ad.gamma_ns((3, 0), br.non_singlet_pids_map["ns-"], 1, nf),
+        ad_us.gamma_ns((3, 0), br.non_singlet_pids_map["ns-"], 1, nf),
         np.zeros(3),
         atol=2e-4,
     )
     assert_allclose(
-        ad.gamma_ns((3, 0), br.non_singlet_pids_map["nsV"], 1, nf),
+        ad_us.gamma_ns((3, 0), br.non_singlet_pids_map["nsV"], 1, nf),
         np.zeros(3),
         atol=8e-4,
     )
     # N3LO
     assert_allclose(
-        ad.gamma_ns((4, 0), br.non_singlet_pids_map["ns-"], 1, nf),
+        ad_us.gamma_ns((4, 0), br.non_singlet_pids_map["ns-"], 1, nf),
         np.zeros(4),
         atol=2e-4,
     )
     # N3LO valence has a spurious pole, need to add a small shift
     assert_allclose(
-        ad.gamma_ns((4, 0), br.non_singlet_pids_map["nsV"], 1 + 1e-6, nf),
+        ad_us.gamma_ns((4, 0), br.non_singlet_pids_map["nsV"], 1 + 1e-6, nf),
         np.zeros(4),
         atol=5e-4,
     )
     assert_raises(
         AssertionError,
         assert_allclose,
-        ad.gamma_ns((4, 0), br.non_singlet_pids_map["ns+"], 1, nf),
+        ad_us.gamma_ns((4, 0), br.non_singlet_pids_map["ns+"], 1, nf),
         np.zeros(4),
     )
